test(suite-native): relocate device connection fixture
