+++ suite-native/device/src/__fixtures__/deviceConnectionFixtures.ts
